// src/enums/unit-type.enum.ts
export enum UnitType {
  ARCH_FIRE_BOILER = 'Arch-fired boiler',
  BUBBLING_FLUIDIZED = 'Bubbling fluidized bed boiler',
  CYCLONE_BOILER = 'Cyclone boiler',
  CELL_BURNER = 'Cell burner boiler',
  COMBINED_CYCLE = 'Combined cycle',
  CIRCULATING_FLUIDIZED = 'Circulating fluidized bed boiler',
  COMBUSTION_TURBINE = 'Combustion turbine',
  DRY_BOTTOM_WALL = 'Dry bottom wall-fired boiler',
  DRY_BOTTOM_TURBO = 'Dry bottom turbo-fired boiler',
  DRY_BOTTOM_VERTICALLY = 'Dry bottom vertically-fired boiler',
  INTERNAL_COMBUSTION = 'Internal combustion engine',
  INTEGRATED_GASIFICATION = 'Integrated gasification combined cycle',
  CEMENT_KLIN = 'Cement Kiln',
  OTHER_BOILER = 'Other boiler',
  OTHER_TURBINE = 'Other turbine',
  PRESSUREIZED_FLUIDIZED = 'Pressurized fluidized bed boiler',
  PROCESS_HEATER = 'Process Heater',
  STOKER = 'Stoker',
  TANGENTIALLY_FIRED = 'Tangentially-fired',
  WET_BOTTOM_WALL = 'Wet bottom wall-fired boiler',
  WET_BOTTOM_TURBO = 'Wet bottom turbo-fired boiler',
  WET_BOTTOM_VERTICALLY = 'Wet bottom vertically-fired boiler',
}