// src/enums/unit-fuel-type.enum.ts
export enum UnitFuelType {
  COAL = 'Coal',
  COAL_REFUSE = 'Coal Refuse',
  DIESEL_OIL = 'Diesel Oil',
  LIQUIFIED_PETROLEUM = 'Liquified Petroleum Gas',
  NATURAL_GAS = 'Natural Gas',
  OTHER_GAS = 'Other Gas',
  OTHER_OIL = 'Other Oil',
  OTHER_SOLID = 'Other Solid Fuel',
  PETROLEUM_COKE = 'Petroleum Coke',
  PIPELINE_NATURAL = 'Pipeline Natural Gas',
  PROCESS_GAS = 'Process Gas',
  PROCESS_SLUDGE = 'Process Sludge',
  REFUSE = 'Refuse',
  RESIDUAL_OIL = 'Residual Oil',
  TIRE_DERIVED = 'Tire Derived Fuel',
  WASTE_LIQUID = 'Waste Liquid',
  WOOD = 'Wood',
}