// src/enums/control-technology.enum.ts
export enum ControlTechnology {
  ADDITIVES_TO_ENHANCE = 'Additives to Enhance PAC and Existing Equipment Performance',
  AMMONIA_INJECTION = 'Ammonia Injection',
  BAGHOUSE = 'Baghouse',
  CATALYST = 'Catalyst (gold, palladium, or other) used to oxidize mercury',
  COMBUSTION_MODIFICATION = 'Combustion Modification/Fuel Reburning',
  CYCLONE = 'Cyclone',
  DRY_LIME_FGD = 'Dry Lime FGD',
  DRY_LOW_NOX = 'Dry Low NOx Burners',
  DRY_SORBENT = 'Dry Sorbent Injection',
  DUAL_ALKALI = 'Dual Alkali',
  ELECTROSTATIC_PRECIPITATOR = 'Electrostatic Precipitator',
  FLUIDIZED_BED_LIMESTONE = 'Fluidized Bed Limestone Injection',
  HALOGENATED_PAC_SORBENT = 'Halogenated PAC Sorbent Injection',
  HYBRID_ESP = 'Hybrid ESP',
  LOW_NOX_DRY_BOTTOM = 'Low NOx Burner Technology (Dry Bottom only)',
  LOW_NOX_CC_OFA = 'Low NOx Burner Technology w/ Closed-coupled OFA',
  LOW_NOX_CC_SEPARATED_OFA = 'Low NOx Burner Technology w/ Closed-coupled/Separated OFA',
  LOW_NOX_OVERFIRE = 'Low NOx Burner Technology w/ Overfire Air',
  LOW_NOX_SEPARATED_OFA = 'Low NOx Burner Technology w/ Separated OFA',
  LOW_NOX_CELL = 'Low NOx Cell Burner',
  MAGNESIUM_OXIDE = 'Magnesium Oxide',
  OTHER = 'Other',
  OTHER_SORBENT = 'Other (Non PAC) Sorbent Injection',
  OVERFIRE_AIR = 'Overfire Air',
  REGENERATIVE_ACTIVATED = 'Regenerative Activated Coke Technology',
  SELECTIVE_CATALYTIC = 'Selective Catalytic Reduction',
  SELECTIVE_NON_CATALYTIC = 'Selective Non-catalytic Reduction',
  SODIUM_BASED = 'Sodium Based',
  STEAM_INJECTION = 'Steam Injection',
  UNTREATED_PAC = 'Untreated PAC Sorbent Injection',
  WATER_INJECTION = 'Water Injection',
  WET_ESP = 'Wet ESP',
  WET_LIME_FGD = 'Wet Lime FGD',
  WET_LIMESTONE = 'Wet Limestone',
  WET_SCRUBBER = 'Wet Scrubber',
}