// src/enums/program.enum.ts
export enum Program {
  ARP = 'ARP',
  CAIRNOX = 'CAIRNOX',
  CAIROS = 'CAIROS',
  CAIRSO2 = 'CAIRSO2',
  CSNOX = 'CSNOX',
  CSNOXOS = 'CSNOXOS',
  CSOSG1 = 'CSOSG1',
  CSOSG2 = 'CSOSG2',
  CSSO2G1 = 'CSSO2G1',
  CSSO2G2 = 'CSSO2G2',
  NBP = 'NBP',
  NHNOX = 'NHNOX',
  NSPS4T = 'NSPS4T',
  OTC = 'OTC',
  RGGI = 'RGGI',
  SIPNOX = 'SIPNOX',
  TXSO2 = 'TXSO2',
}