// src/enums/source-category.enum.ts
export enum SourceCategory {
  AUTOMOTIVE_STAMPINGS = 'Automotive Stampings',
  BULK_INDUSTRIAL_CHEMICAL = 'Bulk Industrial Chemical',
  CEMENT_MANUFACTURING = 'Cement Manufacturing',
  COGENERATION = 'Cogeneration',
  ELECTRIC_UTILITY = 'Electric Utility',
  INDUSTRIAL_BOILER = 'Industrial Boiler',
  INDUSTRIAL_TURBINE = 'Industrial Turbine',
  INSTITUTIONAL = 'Institutional',
  IRON_AND_STEEL = 'Iron & Steel',
  MUNICIPAL_WASTE_COMBUSTOR = 'Municipal Waste Combustor',
  PULP_AND_PAPER_MILL = 'Pulp & Paper Mill',
  PETROLEUM_REFINERY = 'Petroleum Refinery',
  PORTLAND_CEMENT_PLANT = 'Portland Cement Plant',
  SMALL_POWER_PRODUCER = 'Small Power Producer',
  THEME_PARK = 'Theme Park',
}